refactor: calendar 폴더구조 수정
