--- src/entities/date/index.ts
@@ -1,6 +1,0 @@
-export { default as getAllMonthDates } from './lib/get-all-month-dates'
-export { default as getCurrentMonthDates } from './lib/get-current-month-dates'
-
-export { default as ScheduleBadgeFill } from './ui/schedule-badge-fill'
-export { default as ScheduleBadgeMarker } from './ui/schedule-badge-marker'
-export { default as WeekdayHeader } from './ui/weekday-header'
