--- src/features/calendar-interacrtion/index.ts
@@ -1,6 +1,0 @@
-export { default as getDateVariantStates } from './lib/get-date-variant-states'
-
-export { default as CalendarCell } from './ui/calendar-cell'
-export { default as CalendarCellDate } from './ui/calendar-cell-date'
-export { default as CalendarFirstDateLabel } from './ui/calendar-first-date-label'
-export { default as useMonthObserver } from './model/use-month-observer'
