+++ src/shared/ui/calendar/hooks/index.ts
@@ -1,0 +1,2 @@
+export { CalendarProvider, useCalendarContext } from './use-calendar-provider'
+export { default as useMonthObserver } from './use-month-observer'
