+++ src/shared/ui/calendar/lib/get-all-month-dates.ts
@@ -1,4 +1,4 @@
-import getMonthDates from './get-current-month-dates'
+import getCurrentMonthDates from './get-current-month-dates'
 
 /**
  * @description 2025년 1~12월 날짜 배열을 한 번에 모두 리턴하는 함수
@@ -9,8 +9,9 @@ import getMonthDates from './get-current-month-dates'
 
 export default function getAllMonthDates(year: number): Date[][] {
   const allDatesByMonth = Array.from({ length: 12 }, (_, i) => {
-    const date = new Date(year, i, 1)
-    return getMonthDates(date)
+    //12달 짜리 배열
+    const date = new Date(year, i, 1) //각 월의 1일
+    return getCurrentMonthDates(date) //각 월의 전체 날짜
   })
 
   return allDatesByMonth
