+++ src/shared/ui/calendar/lib/go-to-current-month.ts
@@ -1,5 +1,3 @@
-// 📁 src/shared/utils/calendar-utils.ts
-
 import type { MutableRefObject } from 'react'
 
 /**
