+++ src/shared/ui/calendar/ui/weekday-header.tsx
@@ -1,6 +1,6 @@
 import { cva } from 'class-variance-authority'
 
-import Text from '@/shared/ui/text/text'
+import Text from '../../text/text'
 
 /**
  * @description  일~토 요일 표시 헤더,
