--- src/widgets/calendar/index.ts
@@ -1,5 +1,0 @@
-export { default as Calendar } from './ui/calendar'
-export { default as CalendarScrollWrapper } from './ui/calendar-scroll-wrapper'
-export { default as CalendarHeader } from './ui/calendar-header'
-export { default as CalendarMonthView } from './ui/calendar-month-view'
-export { default as goToCurrentMonth } from './lib/go-to-current-month'
